Changed all occurences of 'year' from int64 to uint16.

+++ internal/mortgage/loan.go
@@ -7,10 +7,10 @@ import (
 type Loan struct {
 	Principal  decimal.Decimal
 	AnnualRate decimal.Decimal
-	NumYears   int64
+	NumYears   uint16
 }
 
-func NewLoan(p, r decimal.Decimal, y int64) Loan {
+func NewLoan(p, r decimal.Decimal, y uint16) Loan {
 	return Loan{
 		Principal:  p,
 		AnnualRate: r,
@@ -22,6 +22,6 @@ func (l Loan) MonthlyRate() decimal.Decimal {
 	return l.AnnualRate.Div(decimal.NewFromInt(12))
 }
 
-func (l Loan) NumPeriods() int64 {
+func (l Loan) NumPeriods() uint16 {
 	return l.NumYears * 12
 }
